% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asxdata.R
\docType{data}
\name{asxdata}
\alias{asxdata}
\title{ASX Listed Companies Directory Data}
\format{
\subsection{'asxdata'}{

A data frame with 1943 rows and 5 columns:
\describe{
\item{ASX code}{3 digit code for the listed company}
\item{Company name}{Full Name of the listed company}
\item{GICs industry group}{Industry Classification for the listed company}
\item{Listing data}{Date the company got listed on exchange}
\item{Market Cap}{Market Capitalisation of the listed company as on 20th October 2024}
}
}
}
\source{
<https://www.asx.com.au/markets/trade-our-cash-market/direct
}
\usage{
asxdata
}
\description{
A complete list of companies listed on the Australian Stock Exchange as in October 2024.
}
\keyword{datasets}
