% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozstockpickr-package.R
\docType{package}
\name{ozstockpickr-package}
\alias{ozstockpickr}
\alias{ozstockpickr-package}
\title{ozstockpickr: A Stock Picker for Australian Stock Market}
\description{
The package assists user in making smart decisions while selecting stocks, identifying them by industry and market capitalisation.
}
\keyword{internal}
