% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockdash.R
\name{stockdash}
\alias{stockdash}
\title{Opens Stock Recommendation Dashboard}
\usage{
stockdash()
}
\value{
'DT" returns interactive data table of stock recommendation
}
\description{
Provides the user facility to choose stocks based on Industry & Market Capilisation of the Company
}
\examples{
\dontrun{
run_app()
}
}
